requirements.txt
# 👑 Royal Kingdom Dependencies - The Crown Jewels of Our Platform
# Core Django Framework
asgiref~=3.8.1
Django~=5.2.2
sqlparse~=0.5.1

# Configuration Management
python-decouple~=3.8

# REST API Excellence
djangorestframework~=3.15.2
django-rest-knox~=4.2.0

# WebSocket Magic for Real-Time Features
channels~=4.1.0
channels-redis~=4.2.0

# CORS for Cross-Origin Resource Sharing
django-cors-headers~=4.3.1

# Authentication & Social Login
django-allauth~=0.63.3
python-keycloak~=5.8.1
django-oauth-toolkit~=3.0.1
PyJWT~=2.10.1

# Admin Interface Enhancement
django-jazzmin~=3.0.0

# Developer Tools
django-extensions~=3.2.3
django-browser-reload~=1.13.0

# Database
psycopg2-binary~=2.9.9

# Redis for Caching and Sessions
redis~=5.0.8

# Security & Encryption
cryptography~=42.0.5

# File Upload & Storage
Pillow~=10.4.0

# Utilities
requests~=2.32.3
python-slugify~=8.0.4

# Production Deployment
gunicorn~=23.0.0
whitenoise~=6.6.0

# Development & Testing
pytest~=8.3.2
pytest-django~=4.8.0
factory-boy~=3.3.0
